+++ src/components/adForm/index.tsx
@@ -1,0 +1,1 @@
+export * as AdForm from './AdForm';
